move harness to root folder
